Corrected bugs that caused compile errors for all
I noticed that all the java codes duplicated itself which i suspect to
be merge issues. So I cleaned up the files and corrected all duplicates.

Corrected Multiple “Dex error” - There was duplicated Android Support
Library V4. So i deleted one of them in the project lib file

Deleted the welcome.xml file in the menu folder which was not used in
the code and caused error.